11.1 allows scripts longer than 32K

+++ Jmol/src/org/jmol/viewer/Compiler.java
@@ -147,7 +147,7 @@ class Compiler {
             lnLength *= 2;
           }
           lineNumbers[iCommand] = lineCurrent;
-          lineIndices[iCommand] = (short) ichCurrentCommand;
+          lineIndices[iCommand] = ichCurrentCommand;
           ltoken.setSize(0);
           tokCommand = Token.nada;
           iHaveQuotedString = false;
